Move tasks into tasks package
