+++ src/main/java/duke/command/TodoCommand.java
@@ -1,6 +1,6 @@
 package duke.command;
 
-import duke.Todo;
+import duke.tasks.Todo;
 import duke.ui.TextUi;
 
 /**
