+++ src/main/java/duke/tasks/Event.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.tasks;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
