+++ src/main/java/duke/tasks/Todo.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.tasks;
 
 /**
  * Represents a task that has to be done.
